# src/routes/reports.js
const express = require('express');
const router = express.Router();

// Daily reports endpoint
router.get('/daily', (req, res) => {
  try {
    // TODO: Implement daily reports functionality
    res.json({
      success: true,
      message: 'Daily reports endpoint - implementation pending',
      data: [],
      timestamp: new Date().toISOString()
    });
  } catch (error) {
    console.error('Daily reports error:', error);
    res.status(500).json({
      success: false,
      message: 'Daily reports failed',
      error: error.message
    });
  }
});

// Registration reports endpoint
router.get('/registration', (req, res) => {
  try {
    // TODO: Implement registration reports functionality
    res.json({
      success: true,
      message: 'Registration reports endpoint - implementation pending',
      data: [],
      timestamp: new Date().toISOString()
    });
  } catch (error) {
    console.error('Registration reports error:', error);
    res.status(500).json({
      success: false,
      message: 'Registration reports failed',
      error: error.message
    });
  }
});

// Master reports endpoint
router.get('/master', (req, res) => {
  try {
    // TODO: Implement master reports functionality
    res.json({
      success: true,
      message: 'Master reports endpoint - implementation pending',
      data: [],
      timestamp: new Date().toISOString()
    });
  } catch (error) {
    console.error('Master reports error:', error);
    res.status(500).json({
      success: false,
      message: 'Master reports failed',
      error: error.message
    });
  }
});

module.exports = router;